// extensions/mix-and-match-bundle/generated/api.ts
export type InputMaybe<T> = Maybe<T>;
// ...
export type CartTransformRunResult = {
  /**
   * The ordered list of operations to apply to the cart. For example, you can expand a cart line item to display
   * its [bundled items](https://shopify.dev/docs/apps/build/product-merchandising/bundles), merge
   * multiple cart lines into a single line representing a bundle, and update the presentation of line items
   * in the cart to override their price, title, or image.
   */
  operations: Array<Operation>;
};
// ...
export type LineExpandOperation = {
  /**
   * A [globally-unique ID](https://shopify.dev/docs/api/usage/gids)
   * for a line item in a cart. A cart line represents a single unit of a
   * product variant that a customer intends to purchase.
   */
  cartLineId: Scalars['ID']['input'];
  /**
   * The cart items to expand. Each item in the list is a component of the
   * [bundle](https://shopify.dev/docs/apps/build/product-merchandising/bundles). A
   * bundle is a group of products that are sold together as a single unit.
   */
  expandedCartItems: Array<ExpandedItem>;
  /**
   * The image that replaces the existing image for the group of cart line items.
   * The image must have a publicly accessible URL.
   */
  image?: InputMaybe<ImageInput>;
  /**
   * A change to the original price of a group of items. Price adjustments include discounts or additional charges that
   * affect the final price the customer pays. Price adjustments are often used for promotions, special offers,
   * or any price changes that the customer qualifies for.
   */
  price?: InputMaybe<PriceAdjustment>;
  /**
   * The title that replaces the existing title for the cart line item.
   * If you don't provide a title, then the title of the product variant is used.
   */
  title?: InputMaybe<Scalars['String']['input']>;
};

/**
 * An operation that allows you to override the price, title,
 * and image of a cart line item. Only stores on a
 * [Shopify Plus plan](https://help.shopify.com/manual/intro-to-shopify/pricing-plans/plans-features/shopify-plus-plan)
 * can use apps with `update` operations.
 */
export type LineUpdateOperation = {
  /**
   * A [globally-unique ID](https://shopify.dev/docs/api/usage/gids)
   * for a line item in a cart. A cart line represents a single unit of a
   * product variant that a customer intends to purchase.
   */
  cartLineId: Scalars['ID']['input'];
  /**
   * The image that replaces the existing image for the cart line item.
   * The image must have a publicly accessible URL.
   * This image is visible in checkout only and is not persisted to orders.
   */
  image?: InputMaybe<ImageInput>;
  /**
   * A change to an item's original price. Price adjustments include discounts or additional charges that affect the final
   * price the customer pays. Price adjustments are often used for promotions, special offers, or any price changes
   * for which the customer qualifies.
   */
  price?: InputMaybe<LineUpdateOperationPriceAdjustment>;
  /**
   * The title that replaces the existing title for the cart line item.
   * If you don't provide a title, then the title of the product variant is used.
   */
  title?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type LinesMergeOperation = {
  /**
   * The custom attributes associated with a cart line to store additional information. Cart attributes
   * allow you to collect specific information from customers on the **Cart** page, such as order notes,
   * gift wrapping requests, or custom product details. Attributes are stored as key-value pairs.
   */
  attributes?: InputMaybe<Array<AttributeOutput>>;
  /**
   * The cart items to merge. Each item in the list is a component of the
   * [bundle](https://shopify.dev/docs/apps/build/product-merchandising/bundles).
   */
  cartLines: Array<CartLineInput>;
  /**
   * The image that replaces the existing image for the group of cart line items.
   * The image must have a publicly accessible URL.
   */
  image?: InputMaybe<ImageInput>;
  /**
   * The [globally-unique ID](https://shopify.dev/docs/api/usage/gids)
   * of the product variant that represents the collection of cart line items.
   */
  parentVariantId: Scalars['ID']['input'];
  /**
   * A change to the original price of a group of items. Price adjustments include discounts or additional charges that
   * affect the final price the customer pays. Price adjustments are often used for promotions, special offers,
   * or any price changes that the customer qualifies for.
   */
  price?: InputMaybe<PriceAdjustment>;
  /**
   * The title that replaces the existing title for a group of cart line items.
   * If you don't provide a title, then the title of the parent variant is used.
   */
  title?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type Operation = {
  /**
   * An operation that expands a single cart line item to form a
   * [bundle](https://shopify.dev/docs/apps/build/product-merchandising/bundles)
   * of components.
   */
  lineExpand?: InputMaybe<LineExpandOperation>;
  /**
   * An operation that allows you to override the price, title,
   * and image of a cart line item. Only stores on a
   * [Shopify Plus plan](https://help.shopify.com/manual/intro-to-shopify/pricing-plans/plans-features/shopify-plus-plan)
   * can use apps with `update` operations.
   */
  lineUpdate?: InputMaybe<LineUpdateOperation>;
  /**
   * An operation that merges multiple cart line items into a
   * single line, representing a
   * [bundle](https://shopify.dev/docs/apps/build/product-merchandising/bundles)
   * of components.
   */
  linesMerge?: InputMaybe<LinesMergeOperation>;
};